+++ gitreview_gpt/app.py
@@ -1,13 +1,15 @@
 import os
-import textwrap
 import subprocess
 import json
 import tiktoken
 import argparse
 import re
+import gitreview_gpt.prompt as prompt
+import gitreview_gpt.formatter as formatter
 from yaspin import yaspin
 
 
+# Return the code changes as a git diff
 def get_git_diff(staged, branch):
     # Run git diff command and capture the output
     if not branch:
@@ -20,138 +22,20 @@ def get_git_diff(staged, branch):
     return git_diff.stdout
 
 
-# Format the git diff into a format that can be used by the GPT-3 API
-# - add line numbers to the diff
-# - split the diff into chunks per file
-def format_git_diff(diff_text):
-    diff_formatted = ""
-    file_chunks = {}
-    file_names = {}
-
-    # Split git diff into chunks with separator +++ line inclusive, the line with the filename
-    pattern = r"(?=^(\+\+\+).*$)"
-    parent_chunks = re.split(r"\n\+{3,}\s", diff_text, re.MULTILINE)
-    for j, chunk in enumerate(parent_chunks, 0):
-        # skip first chunk (it's the head info)
-        if j == 0:
-            continue
-
-        # Remove git --diff section
-        chunk = re.sub(r"(diff --git.*\n)(.*\n)", lambda match: match.group(1), chunk)
-        chunk = re.sub(r"^diff --git.*\n", "", chunk, flags=re.MULTILINE)
-
-        # Split chunk into chunks with separator @@ -n,n +n,n @@ inclusive, the changes in the file
-        pattern = r"(?=@@ -\d+,\d+ \+\d+,\d+ @@)"
-        chunks = re.split(pattern, chunk)
-        for i, chunk in enumerate(chunks, 1):
-            # skip first chunk (it's the file name)
-            if i == 1:
-                diff_formatted += chunk
-                file_chunks[j] = chunk
-                file_names[j - 1] = chunk.rstrip("\n").rsplit("/", 1)[-1]
-                continue
-
-            # Extract the line numbers from the changes pattern
-            pattern = r"@@ -(\d+),(\d+) \+(\d+),(\d+) @@"
-            match = re.findall(pattern, chunk)
-            chunk_dividers = []
-            for m in match:
-                chunk_dividers.append(
-                    {
-                        "original_start_line": int(m[0]),
-                        "original_end_line": int(m[1]),
-                        "new_start_line": int(m[2]),
-                        "new_end_line": int(m[3]),
-                    }
-                )
-            line_counter = -1 + chunk_dividers[0]["new_start_line"]
-
-            chunk_formatted = ""
-            for line in chunk.splitlines():
-                if line.startswith("@@ -"):
-                    diff_formatted += line + "\n"
-                    chunk_formatted += line + "\n"
-                    continue
-                if line.startswith("---"):
-                    continue
-                if line.startswith("-"):
-                    continue
-                else:
-                    line_counter += 1
-
-                new_line = str(line_counter) + " " + line + "\n"
-                diff_formatted += new_line
-                chunk_formatted += new_line
-
-            file_chunks[j] += chunk_formatted
-
-    # TODO remove lines with pattern @@ -n,n +n,n @@ from diff_formatted and file_chunks
-    return diff_formatted, file_chunks, file_names
-
-
-# return the number of tokens in a string
+# Return the number of tokens in a string
 def count_tokens(text):
     encoding = tiktoken.encoding_for_model("gpt-3.5-turbo")
     tokenized = encoding.encode(text)
     return len(tokenized)
 
 
-def get_commit_message_prompt(diff_text):
-    return {
-        "model": "gpt-3.5-turbo",
-        "max_tokens": 2000,
-        "temperature": 0.5,
-        "n": 1,
-        "stop": None,
-        "messages": [
-            {
-                "role": "user",
-                "content": "Here are my code changes. Could you please create a short and precise commit message for me?",
-            },
-            {
-                "role": "assistant",
-                "content": "Sure! I'll be happy to help. Please share the code changes you made.",
-            },
-            {
-                "role": "user",
-                "content": diff_text,
-            },
-        ],
-    }
-
-
-def get_review_prompt(diff_text):
-    return {
-        "model": "gpt-3.5-turbo",
-        "max_tokens": 2000,
-        "temperature": 0.5,
-        "n": 1,
-        "stop": None,
-        "messages": [
-            {
-                "role": "user",
-                "content": "You are a code reviewer. You should review my code changes and provide feedback. "
-                + "Provide feeback on how to improve the code. Provide feedback on potential bugs. "
-                + "You will get my changes with line numbers at the start of each line. "
-                + "Provide feedback in the following format: $$$filename:line_numbers:$$$feedback. For example, $$$main.py:10:$$$This line of code is not required.",
-            },
-            {
-                "role": "assistant",
-                "content": "Sure! Please share the code changes you made.",
-            },
-            {
-                "role": "user",
-                "content": diff_text,
-            },
-        ],
-    }
-
-
+# Send request with prompt as payload to openai completions api
 def send_request(api_key, payload, spinner_text):
     # Convert payload to JSON string
     payload_json = json.dumps(payload).replace("'", r"'\''")
+    # TODO also replace " ?
 
-    # create spinner
+    # Create spinner
     spinner = yaspin()
     spinner.text = spinner_text
     spinner.start()
@@ -164,12 +48,13 @@ def send_request(api_key, payload, spinner_text):
         curl_command, shell=True, capture_output=True, text=True
     )
 
-    # stop spinner
+    # Stop spinner
     spinner.stop()
 
     # Process the response
     if curl_output.returncode == 0:
         json_response = json.loads(curl_output.stdout)
+        # TODO check if 'choices' available
         reviewSummary = (
             json_response["choices"][0]["message"]["content"]
             .encode()
@@ -180,49 +65,33 @@ def send_request(api_key, payload, spinner_text):
         return f"Error: {curl_output.stderr.strip()}"
 
 
-def draw_box(filename, feedback_lines):
-    max_length = max(
-        len(f"{line['line']}: {line['feedback']}") for line in feedback_lines
-    )
-    max_length = os.get_terminal_size()[0] - 2
-    border = "╭" + "─" * (max_length) + "╮"
-    bottom_border = "│" + "─" * (max_length) + "│"
-    filename_line = "│ " + filename.ljust(max_length - 1) + "│"
-    result = [border, filename_line, bottom_border]
-
-    for entry in feedback_lines:
-        line_string = f"Line {entry['line']}: {entry['feedback']}"
-        if len(line_string) > max_length - 2:
-            wrapped_lines = textwrap.wrap(line_string, width=max_length - 2)
-            result.append("│ " + wrapped_lines[0].ljust(max_length - 2) + " │")
-            for wrapped_line in wrapped_lines[1:]:
-                result.append("│ " + wrapped_line.ljust(max_length - 2) + " │")
+# Retrieve review from openai completions api
+# Process response and send repair request if json has invalid format
+def request_review(api_key, code_to_review):
+    payload = prompt.get_review_prompt(code_to_review)
+    review_result = send_request(api_key, payload, "Reviewing...")
+    print(review_result + "\n")
+    try:
+        get_review_output_from_response_json(review_result)
+    except ValueError as e:
+        print("Review result has bad format. It will be repaired.")
+        payload = prompt.get_review_repair_prompt(review_result, e)
+        review_result = send_request(api_key, payload, "Repairing...")
+        print(review_result + "\n")
+        pattern = r"json\s*({[^`]+})\s*"
+        match = re.search(pattern, review_result, re.DOTALL)
+        if match:
+            get_review_output_from_response_json(match.group(1))
         else:
-            result.append("│ " + line_string.ljust(max_length - 2) + " │")
-
-    result.append("╰" + "─" * (max_length) + "╯")
-    return "\n".join(result)
+            get_review_output_from_response_json(review_result)
 
 
-def get_review_output_text(review_result):
-    parsed_feedback = {}
-
-    pattern = r"\$\$\$(.*?):(\d+).*?\$\$\$(.*)"
-    matches = re.findall(pattern, review_result)
-
-    for match in matches:
-        filename = match[0]
-        line_number = match[1]
-        feedback = match[2].strip()
-
-        if filename not in parsed_feedback:
-            parsed_feedback[filename] = []
-
-        parsed_feedback[filename].append({"line": line_number, "feedback": feedback})
-
-    # Print the parsed feedback
-    for filename, feedback_list in parsed_feedback.items():
-        print(draw_box(filename, feedback_list))
+# Process response json and draw output to console
+def get_review_output_from_response_json(feedback_json):
+    feedback = json.loads(feedback_json)
+    print("✨ Review Result ✨")
+    for file in feedback:
+        print(formatter.draw_box(file, feedback[file]))
 
 
 def run():
@@ -267,7 +136,7 @@ def run():
             print("No git changes.")
         exit()
 
-    formatted_diff, diff_file_chunks, file_names = format_git_diff(diff_text)
+    formatted_diff, diff_file_chunks, file_names = formatter.format_git_diff(diff_text)
 
     token_count = count_tokens(formatted_diff)
 
@@ -301,24 +170,16 @@ def run():
                         "TODO: token count exceeds 1500. Split file chunks into chunk of changes"
                     )
                     exit()
-                prompt = get_review_prompt(value)
-                review_result = send_request(api_key, prompt, "Reviewing...")
-                # TODO send repair request if review result has bad format
-                print("✨ Review Result ✨")
-                get_review_output_text(review_result)
+                request_review(api_key, value)
 
         # Review the changes in one request
         else:
-            prompt = get_review_prompt(formatted_diff)
-            review_result = send_request(api_key, prompt, "Reviewing...")
-            # TODO send repair request if review result has bad format
-            print("✨ Review Result ✨")
-            get_review_output_text(review_result)
+            request_review(api_key, formatted_diff)
 
     # Create a commit message using OpenAI API
     if args.action == "commit":
-        prompt = get_commit_message_prompt(formatted_diff)
-        review_result = send_request(api_key, prompt, "Creating commit message...")
+        payload = prompt.get_commit_message_prompt(formatted_diff)
+        review_result = send_request(api_key, payload, "Creating commit message...")
         print("✨ Commit Message ✨")
         print(review_result)
         print("Do you want to commit the changes? (y/n)")
